Moved code out of clients for better testability.
