+++ test/test_helper.rb
@@ -15,7 +15,7 @@ if ENV["COVERAGE"]
   # for coverage we need to load all ruby files
   src_location = File.expand_path("../../src", __FILE__)
   # note that clients/ are excluded because they run too eagerly by design
-  Dir["#{src_location}/{include,modules}/**/*.rb"].each do |f|
+  Dir["#{src_location}/{include,lib,modules}/**/*.rb"].each do |f|
     require_relative f
   end
